Refactor / clean up a few random things
