+++ include/graph_generator.h
@@ -21,20 +21,6 @@ namespace detail {
 	class abstract_command;
 	class horizon_command;
 
-	// TODO: Move to utility header..?
-	// Implementation from Boost.ContainerHash, licensed under the Boost Software License, Version 1.0.
-	inline void hash_combine(std::size_t& seed, std::size_t value) { seed ^= value + 0x9e3779b9 + (seed << 6) + (seed >> 2); }
-
-	struct pair_hash {
-		template <typename U, typename V>
-		std::size_t operator()(const std::pair<U, V>& p) const {
-			std::size_t seed = 0;
-			hash_combine(seed, std::hash<U>{}(p.first));
-			hash_combine(seed, std::hash<V>{}(p.second));
-			return seed;
-		}
-	};
-
 	class graph_generator {
 		friend struct graph_generator_testspy;
