+++ include/utils.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <cstdint>
+#include <functional>
 #include <type_traits>
 #include <variant>
 
@@ -27,5 +28,18 @@ template <typename Variant, typename... Arms>
 decltype(auto) match(Variant&& v, Arms&&... arms) {
 	return std::visit(overload{std::forward<Arms>(arms)...}, std::forward<Variant>(v));
 }
+
+// Implementation from Boost.ContainerHash, licensed under the Boost Software License, Version 1.0.
+inline void hash_combine(std::size_t& seed, std::size_t value) { seed ^= value + 0x9e3779b9 + (seed << 6) + (seed >> 2); }
+
+struct pair_hash {
+	template <typename U, typename V>
+	std::size_t operator()(const std::pair<U, V>& p) const {
+		std::size_t seed = 0;
+		hash_combine(seed, std::hash<U>{}(p.first));
+		hash_combine(seed, std::hash<V>{}(p.second));
+		return seed;
+	}
+};
 
 } // namespace celerity::detail::utils
