+++ test/graph_generation_tests.cc
@@ -569,7 +569,7 @@ namespace detail {
 		const range<1> node_range{num_nodes};
 
 		// TODO placeholder: complete with dependency types for other side effect orders
-		const auto expected_dependencies = std::unordered_map<std::pair<order, order>, std::optional<dependency_kind>, pair_hash>{
+		const auto expected_dependencies = std::unordered_map<std::pair<order, order>, std::optional<dependency_kind>, utils::pair_hash>{
 		    {{order::sequential, order::sequential}, dependency_kind::true_dep},
 		};
